backend/requirements.txt
# Backend Dependencies for Refinement & Validation

# Existing dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
aiohttp==3.9.1
Pillow==10.1.0

# New dependencies for Steps 5 & 6

# Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.22.0

# Face Recognition & Validation
deepface==0.0.79
tf-keras==2.15.0  # Required by DeepFace
retina-face==0.0.13

# Image Quality Metrics
brisque==0.0.4

# EXIF Metadata
piexif==1.1.3

# Image Format Support
pillow-heif==0.13.1  # HEIF format
Pillow-SIMD==9.5.0.post0  # Faster image operations (optional)

# Super-Resolution (optional, large models)
# realesrgan==0.3.0  # Uncomment for production
# basicsr==1.4.2

# Utilities
python-dateutil==2.8.2
